* include/openbabel/atom.h, include/openbabel/mol.h: Fix   PR#1626789 with large PDB files. Fixes regression to 2006-09-25   changes supporting >2^16 atoms or bonds.
  * src/mol.cpp(EndModify): Ensure idx is unsigned (to match atom indexes).

  * include/openbabel/chains.h, src/chains.cpp: Make sure atomids is
  int, not short. Atom IDs can go up to a full unsigned int.

+++ include/openbabel/atom.h
@@ -89,7 +89,7 @@ namespace OpenBabel
       OBMol                        *_parent;    //!< parent molecule (if any)
       std::vector<OBBond*>          _vbond;     //!< bonds to this atom -- assumed to be one of the endpoints
 
-      unsigned short                _cidx;      //!< index into coordinate array
+      unsigned int                  _cidx;      //!< index into coordinate array
       unsigned short                _hyb;       //!< hybridization
       unsigned short                _flags;     //!< bitwise flags (e.g. aromaticity)
       double                        _pcharge;   //!< partial charge
